Modularized client-side/chat code for readability
